Init repository (#1)
* Init repository

Added an implementation of a MaxScale CDC protocol client.

* Declared the different event types of the CDC protocol

- Created new structs to represent DDL and DML events
- Added new integration tests to test the REQUEST-DATA step of the protocol
- Added a monitor service to the MaxScale configuration

* Fix tests

- Added an healthcheck on the MaxScale service checking for the existence of the generated avro file. Otherwise the test-runner might start before it get generated and fails.
- Changed the condition used to differentiate DDL from DML events. Old one could consider DML events as DDL ones.

* Added the table data to the DML event

- Now unmarshal the table data into a separate fields
- Added a new `TableData` field to the expected DML event in the tests
- Cleaned a bit the code

* Improved tests

Added a test to verify if the client starts reading at the given GTID.

* Updated API

Modified API in order to be able to set read/write timeouts only once on the client.

* Code cleaning

- Updated tests names
- Added comments

* Updated API

Replaced the MaxScale CDC client by a Connection struct so the object is always in a usable state.

* Added go-cmp package

Improved the readability of differences when comparing structs in tests.

* Updated API

Merged all methods together to avoid missusage of the API.

* Fixed error handling

Modified the code in order to return an error directly if the client cannot read the schema on the first read. This aims to facilitate the error handling for the API consumer.

* Added a new method

Added a GTID method to retrieve the GTID of the event.

* Updated RequestData workflow

- Implemented graceful shutdown(might have to double check later)
- Now do not consider missing .avro file as a fatal error

* Fixed the event decoding workflow

* Improved event decoding

After looking at https://github.com/mariadb-corporation/MaxScale/blob/22.08/server/modules/protocol/examples/cdc.py#L41 it seems that every event/error are written on a single line. Therefore I modified the event decoding workflow to scan the socket line by line.

* Added option

Added a new option to inject logger into the client.

* Dockerfile and Makefile update

- Removed unused `build` rule from the Makefile
- Renamed the Dockerfile to follow the naming used in other projects

* Fixed the graceful shutdown

Removed the context to avoid confusion on how to stop the client to request data to the MaxScale CDC listener.

* Removed dependency

Removed the `uuid` package from the dependency as it was unneeded.

* Modified DML event struct

Replaced the `TableData` field by the `Raw` field which is the raw bytes representing the event. It is then easier to work with for this package's API consumer.

* Updated docker-compose

Removed unused port for clarity.

* Updated naming and fixed integration tests

- Removed the "cdc" prefix everywhere since we changed the package name by "cdc"
- Fixed the integration tests by removing the comparison which involved the timestamp returned by MaxScale

* Added new method

Implemented a new `TableData` method on the DML event which unmarshal the raw bytes into a map[string]interface{} without the fields not related to the table data.

* Replaced escaped double quote by %q verbe in `fmt.Sprintf` calls.