Fixed raw event (#2)
* Fixed raw event

Apparently the `Bytes` method from the `bufio.Scanner` does not make any allocation and reuse the same byte slice as a ring buffer. Therefore we need to make a copy to otherwise it'll be altered on the next `Scan` call.
See https://pkg.go.dev/bufio#Scanner.Bytes

* .gitignore update

Commented the `*.text` ignore directive to allow tracking of the test Dockerfile.

* Code documentation

Added more comments on type exposed by this package.